requirements.txt
streamlit>=1.53.0
sqlalchemy>=2.0.0
alembic>=1.14.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
openpyxl>=3.1.0
scikit-learn>=1.6.0
requests>=2.31.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
